+++ Sources/SuperwallKit/Superwall.swift
@@ -273,7 +273,7 @@ public final class Superwall: NSObject, ObservableObject {
       _ = await (fetchConfig, configureIdentity)
 
       await track(
-        InternalSuperwallPlacement.ConfigAttributes(
+        InternalSuperwallEvent.ConfigAttributes(
           options: dependencyContainer.configManager.options,
           hasExternalPurchaseController: purchaseController != nil,
           hasDelegate: delegate != nil
@@ -310,29 +310,32 @@ public final class Superwall: NSObject, ObservableObject {
     $subscriptionStatus
       .removeDuplicates()
       .dropFirst()
-      .scan((previous: subscriptionStatus, current: subscriptionStatus)) { previousPair, newStatus in
+      .scan((previous: subscriptionStatus, current: subscriptionStatus)) {
+        previousPair, newStatus in
         // Shift the current value to previous, and set the new status as the current value
         (previous: previousPair.current, current: newStatus)
       }
       .receive(on: DispatchQueue.main)
-      .subscribe(Subscribers.Sink(
-        receiveCompletion: { _ in },
-        receiveValue: { [weak self] statusPair in
-          guard let self = self else {
-            return
-          }
-          let oldStatus = statusPair.previous
-          let newStatus = statusPair.current
+      .subscribe(
+        Subscribers.Sink(
+          receiveCompletion: { _ in },
+          receiveValue: { [weak self] statusPair in
+            guard let self = self else {
+              return
+            }
+            let oldStatus = statusPair.previous
+            let newStatus = statusPair.current
 
-          self.dependencyContainer.storage.save(newStatus, forType: SubscriptionStatusKey.self)
+            self.dependencyContainer.storage.save(newStatus, forType: SubscriptionStatusKey.self)
 
-          Task {
-            await self.dependencyContainer.delegateAdapter.subscriptionStatusDidChange(from: oldStatus, to: newStatus)
-            let event = InternalSuperwallPlacement.SubscriptionStatusDidChange(status: newStatus)
-            await Superwall.shared.track(event)
+            Task {
+              await self.dependencyContainer.delegateAdapter.subscriptionStatusDidChange(
+                from: oldStatus, to: newStatus)
+              let event = InternalSuperwallEvent.SubscriptionStatusDidChange(status: newStatus)
+              await Superwall.shared.track(event)
+            }
           }
-        }
-      ))
+        ))
   }
 
   /// Sets ``subscriptionStatus`` to an`unknown` state.
@@ -448,14 +451,14 @@ public final class Superwall: NSObject, ObservableObject {
 
   /// Confirms all experiment assignments and returns them in an array.
   ///
-  /// This tracks ``SuperwallPlacement/confirmAllAssignments`` in the delegate.
+  /// This tracks ``SuperwallEvent/confirmAllAssignments`` in the delegate.
   ///
   /// Note that the assignments may be different when a placement is registered due to changes
   /// in user, placement, or device parameters used in audience filters.
   ///
   /// - Returns: An array of ``Assignment`` objects.
   public func confirmAllAssignments() async -> [Assignment] {
-    let confirmAllAssignments = InternalSuperwallPlacement.ConfirmAllAssignments()
+    let confirmAllAssignments = InternalSuperwallEvent.ConfirmAllAssignments()
     await track(confirmAllAssignments)
 
     guard let triggers = dependencyContainer.configManager.config?.triggers else {
@@ -487,7 +490,7 @@ public final class Superwall: NSObject, ObservableObject {
 
   /// Confirms all experiment assignments and returns them in an array.
   ///
-  /// This tracks ``SuperwallPlacement/confirmAllAssignments`` in the delegate.
+  /// This tracks ``SuperwallEvent/confirmAllAssignments`` in the delegate.
   ///
   /// Note that the assignments may be different when a placement is registered due to changes
   /// in user, placement, or device parameters used in audience filters.
@@ -618,7 +621,7 @@ public final class Superwall: NSObject, ObservableObject {
 
     Task {
       let deviceAttributes = await dependencyContainer.makeSessionDeviceAttributes()
-      let deviceAttributesPlacement = InternalSuperwallPlacement.DeviceAttributes(
+      let deviceAttributesPlacement = InternalSuperwallEvent.DeviceAttributes(
         deviceAttributes: deviceAttributes)
       await track(deviceAttributesPlacement)
     }
@@ -631,7 +634,7 @@ public final class Superwall: NSObject, ObservableObject {
     Task {
       let deviceAttributes = await dependencyContainer.makeSessionDeviceAttributes()
       await Superwall.shared.track(
-        InternalSuperwallPlacement.DeviceAttributes(deviceAttributes: deviceAttributes)
+        InternalSuperwallEvent.DeviceAttributes(deviceAttributes: deviceAttributes)
       )
     }
   }
@@ -650,7 +653,7 @@ public final class Superwall: NSObject, ObservableObject {
   @discardableResult
   public func handleDeepLink(_ url: URL) -> Bool {
     Task {
-      await track(InternalSuperwallPlacement.DeepLink(url: url))
+      await track(InternalSuperwallEvent.DeepLink(url: url))
     }
     return dependencyContainer.debugManager.handle(deepLinkUrl: url)
   }
@@ -687,7 +690,7 @@ public final class Superwall: NSObject, ObservableObject {
     presentationItems.reset()
     dependencyContainer.configManager.reset()
     Task {
-      await Superwall.shared.track(InternalSuperwallPlacement.Reset())
+      await Superwall.shared.track(InternalSuperwallEvent.Reset())
 
       #if os(iOS) || os(macOS) || os(visionOS)
         if #available(iOS 14.3, macOS 11.1, macCatalyst 14.3, *) {
@@ -981,7 +984,7 @@ extension Superwall: PaywallViewControllerEventDelegate {
     case let .customPlacement(name: name, params: params):
       Task {
         let paramsDict = params.dictionaryValue
-        let customPlacement = InternalSuperwallPlacement.CustomPlacement(
+        let customPlacement = InternalSuperwallEvent.CustomPlacement(
           paywallInfo: paywallViewController.info,
           name: name,
           params: paramsDict